refactor(dto): reorganize User request DTO into specific subpackages
Move User related DTO into subpackage dto.request.user.
